Reformatted and rebuilt the heightmap rendering.
Now interpolates between the ocean and beach color based on water depth.  This provides a look much closer to SL's.

Disabled the self-shadowing code as it was causing large blocks of terrain to go dark and look weird.

Rebuilt the water-terrain logic to make things cleaner and be able to get underwater detail when the shadowing code gets fixed.