Switched out from bandying around DirectBitmap instances to the basic Bitmap instances.
DirectBitmap only comes into its own when being drawn upon with lots of little operations.  When it's just a picture holder that might have some bulk ops done to it it doesn't need to have the extra luggage.

Plus a DirectBitmap doesn't have all the nice ctors that a raw Bitmap does, and my existing tools shoudl be accepting of things doen that way.