Restructured database access to reduce memory consumption
When tested against a large grid Anax went insane on memory consumption: trying to keep the grid's data in memory was just too much! :P  Of course the tradeoff was performace.  How much? Dunno, haven't profiled yet.

So I taught prims and terrain how to get themselves from the DB, and made the regions gatekeepers for the data instead of containers of the data.  Now a batch of prims or a terrain is something that comes, gets processed, and then is up for GC.

This also means that the RDBMap became a lot simpler, and much more focused on being a map access tool.

I did break servermode more, but that section isn't functional anyway ATM.